refactor: moved timetable building code to a separate file
