+++ wimp/__init__.py
@@ -1,3 +1,4 @@
 from .fetch import *
 from .parse import *
+from .timetable import *
 from .constants import *
